Relocate header.js
